backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
browser-use-sdk>=3.0.0
anthropic>=0.39.0
pydantic>=2.0.0
python-dotenv>=1.0.0
PyJWT[crypto]>=2.8.0
google-genai>=1.0.0
